@@ -1,7 +1,7 @@
 package leavesc.hello.filetransfer;
 
 import android.annotation.SuppressLint;
-import android.support.v7.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatActivity;
 import android.widget.Toast;
 
 import leavesc.hello.filetransfer.common.LoadingDialog;
