@@ -1,8 +1,8 @@
 package leavesc.hello.filetransfer.adapter;
 
 import android.net.wifi.p2p.WifiP2pDevice;
-import android.support.annotation.NonNull;
-import android.support.v7.widget.RecyclerView;
+import androidx.annotation.NonNull;
+import androidx.recyclerview.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
