@@ -1,6 +1,6 @@
 package leavesc.hello.filetransfer.model;
 
-import android.support.annotation.NonNull;
+import androidx.annotation.NonNull;
 
 import java.io.Serializable;
 
